pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.14"

dependencies = [
    "livekit-agents[groq,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-deepgram",
    "livekit-plugins-openai",
    "livekit-plugins-cartesia",
    "python-dotenv",
    "langchain-groq>=1.1.2",
    "langchain-core>=1.2.13",
    "chromadb>=1.5.0",
    "sentence-transformers>=5.2.3",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "langgraph>=1.0.8",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
